+++ src/index.js
@@ -16,6 +16,7 @@ import secrets from '@adobe/helix-shared-secrets';
 import bodyData from '@adobe/helix-shared-body-data';
 import dataAccess from '@adobe/spacecat-shared-data-access';
 import {
+  badRequest,
   internalServerError,
   noContent,
   notFound,
@@ -43,6 +44,12 @@ import trigger from './controllers/trigger.js';
 import { App as SlackApp } from './utils/slack/bolt.cjs';
 import ConfigurationController from './controllers/configuration.js';
 import FulfillmentController from './controllers/event/fulfillment.js';
+import ImportController from './controllers/import.js';
+import { s3ClientWrapper } from './support/s3.js';
+
+const uuidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+
+const isValidUUIDV4 = (uuid) => uuidRegex.test(uuid);
 
 export function enrichPathInfo(fn) { // export for testing
   return async (request, context) => {
@@ -95,12 +102,20 @@ async function run(request, context) {
       SlackController(SlackApp),
       trigger,
       FulfillmentController(context),
+      ImportController(context),
     );
 
     const routeMatch = matchPath(method, suffix, routeHandlers);
 
     if (routeMatch) {
       const { handler, params } = routeMatch;
+      //
+      if (params.siteId && !isValidUUIDV4(params.siteId)) {
+        return badRequest('Site Id is invalid. Please provide a valid UUID.');
+      }
+      if (params.organizationId && (!isValidUUIDV4(params.organizationId) && params.organizationId !== 'default')) {
+        return badRequest('Organization Id is invalid. Please provide a valid UUID.');
+      }
       context.params = params;
 
       return await handler(context);
@@ -124,6 +139,7 @@ export const main = wrap(run)
   .with(enrichPathInfo)
   .with(bodyData)
   .with(sqs)
+  .with(s3ClientWrapper)
   .with(imsClientWrapper)
   .with(elevatedSlackClientWrapper, { slackTarget: WORKSPACE_EXTERNAL })
   .with(secrets, { name: resolveSecretsName })