+++ src/support/slack/commands.js
@@ -17,6 +17,7 @@ import getSite from './commands/get-site.js';
 import getSites from './commands/get-sites.js';
 import martechImpact from './commands/martech-impact.js';
 import runAudit from './commands/run-audit.js';
+import runImport from './commands/run-import.js';
 import setLiveStatus from './commands/set-live-status.js';
 import help from './commands/help.js';
 import bulkUpdateAuditConfigs from './commands/bulk-update-audits.js';
@@ -35,6 +36,7 @@ export default (context) => [
   getSites(context),
   martechImpact(context),
   runAudit(context),
+  runImport(context),
   setLiveStatus(context),
   bulkUpdateAuditConfigs(context),
   help(context),